v 2.3.2 19 december 08
git-svn-id: svn+ssh://tictacserver.gel.usherbrooke.ca/Users/tictac/svnServer/developments/fullDiamant/diamant/trunk@2757 bc377d3f-2bd2-4e0d-9bbf-68e52cfa64df

+++ src/dInternal/dData/DLoadData.java
@@ -322,7 +322,7 @@ public class DLoadData {
 
 		String dataloaded = new String(filterBadChars(fileName));
 		StringTokenizer dataTokens;
-		long linePosition = 0;
+		int linePosition = 0;
 		DataExchange de;
 
 		dataTokens = new StringTokenizer(dataloaded, DConst.SAVE_SEPARATOR);
